+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/work/WorkChore.kt
@@ -22,7 +22,7 @@ class WorkChore(
     deleted: Boolean? = false,
     userId: Int,
     finished: Boolean? = false,
-    color: Color? = Color.Companion.Blue,
+    color: Color? = Color.Blue,
 
     val project: String,
     val clientName: String? = null,